try to fix warnings

+++ drivers/power/supply/ti/ln8000_charger.c
@@ -1392,7 +1392,7 @@ static const struct regmap_config ln8000_regmap_config = {
     .max_register	= LN8000_REG_MAX,
 };
 
-static int ln8000_get_dev_role(struct i2c_client *client)
+static uintptr_t ln8000_get_dev_role(struct i2c_client *client)
 {
     const struct of_device_id *of_id;
 
@@ -1404,7 +1404,7 @@ static int ln8000_get_dev_role(struct i2c_client *client)
 
     dev_info(&client->dev,"%s: matched to %s\n", __func__, of_id->compatible);
 
-    return (int)of_id->data;
+    return (uintptr_t)of_id->data;
 }
 
 static int ln8000_parse_dt(struct ln8000_info *info)
